requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
boto3==1.37.20
gunicorn==21.2.0 